Lab5_docker_docker_compose_makefile/frontend/requirements.txt
streamlit==1.32.0
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.0